pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "farswarm"
version = "0.1.0"
description = "Brain-encoded swarm social simulation engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Samuel Zhao" }]
keywords = [
    "neuroscience",
    "multi-agent",
    "simulation",
    "brain-encoding",
    "prediction",
    "social-simulation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "typer>=0.12",
    "rich>=13.7",
    "openai>=1.12",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "plotly>=5.20",
    "pandas>=2.2",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
oasis = [
    "camel-oasis>=0.2.5",
    "camel-ai>=0.2.78",
]
tribe = [
    "tribev2",
    "torch>=2.5.1,<2.7",
    "torchvision>=0.20,<0.22",
]
embeddings = [
    "sentence-transformers>=3.0",
]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.3",
    "mypy>=1.9",
]

[project.scripts]
farswarm = "farswarm.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
